refactor: legg klasser og webclient i egne filer
