--- .env.local
@@ -1,9 +1,0 @@
-ISSUER_URI="https://kv-ros-backstage-245zlcbrnq-lz.a.run.app/api/auth"
-GITHUB_APP_ID="827078"
-GITHUB_INSTALLATION_ID="48036152"
-GITHUB_PRIVATE_KEY_BASE64_ENCODED="<ligger i projects/457384642040/secrets/GITHUB_PRIVATE_KEY_BASE64_ENCODED_DEV>"
-RISC_FOLDER_PATH=".security/risc"
-JSON_SCHEMA_BASE_URL="https://kartverket.github.io/backstage-plugin-risk-scorecard-backend"
-ENTRA_ID_TENANT_ID="c6ea0b0b-e867-46d4-a70b-4c05887e7f6f"
-FILENAME_PREFIX="risc"
-FILENAME_POSTFIX="risc"
