+++ src/main/kotlin/no/risc/exception/exceptions/RiScNotValidException.kt
@@ -5,3 +5,13 @@ data class RiScNotValidException(
     val riScId: String,
     val validationError: String,
 ) : Exception()
+
+data class RiScCouldNotCreateNewException(
+    override val message: String,
+    val riScId: String,
+) : Exception()
+
+data class RiScCouldNotUpdateException(
+    override val message: String,
+    val riScId: String,
+) : Exception()
