+++ src/main/kotlin/no/risc/github/models/ContentStatus.kt
@@ -1,0 +1,11 @@
+package no.risc.github.models
+
+enum class UpdateContentStatus {
+    Success,
+    Failure,
+}
+
+enum class CreateContentStatus {
+    Success,
+    Failure,
+}
